moves `process_messages()` method to scan_task.rs
